Removing OAuth2Authentiction class from rest_framework_oauth library
The reset_framework_oauth library(called djangorestframework-oauth on pypi) has been deprecated
for quite a while. This removes any references to library in code.

It was verified using new relic metrics that the class OAuth2Authentication was not actively used
to authenticate any user. If it is found that the class is necessary, add back library in this repo
and in edx-platform and make sure to pull from edx fork on library rather than pull from pypi.