Add dry_run option to bulk-create functions
The dry run will log what orgs/org-courses would be created,
but it does not apply the changes.

Also, improves logging a bit so that we actually log the orgs
and org-courses that will be *re*-activated. Currently,
we log those that are *existing*, which is a superset of those
that will be re-activated.

For use in backfill_orgs_and_org_courses Studio management
command.

Bump version from 6.1.0 to 6.2.0

TNL-7646