Stop checking for FEATURES['ORGANIZATIONS_APP'] setting
We originally checked the deprecated
`FEATURES['ORGANIZATIONS_APP']` setting
as an alternative to the new `ORGANIZATIONS_AUTOCREATE`
settting, but since we're going to
give `ORGANIZATIONS_AUTOCREATE` a default value in
edx-platform anyway, the fallback check will never
even be reached. Having the fallback check in the
edx-organizations codebase is more misleading than
anything else.

This is effectively a no-op change.

Bump version to 6.4.0

TNL-7646