Allow organizations and linkages to be bulk-added as inactive
Add a new optional flag `activate` to both:
  * `api.bulk_add_organizations` and
  * `api.bulk_add_organization_courses`,
defaulting to `True`, which is the current behavior.

With `activate=False`, those functions will set
the `active` flag on newly created objects to
`False`. Furthermore, they will not re-
activate any existing objects.

The purpose of this is to allow us to backfill
the Organization and OrganizationCourse tables
on production instances without introducing any
new active data. We intend to expose this flag
as a command-line flag in the edx-platform
`bulk_add_orgs_and_org_courses` command.

Bump version to 6.6.0

TNL-7774