Allow inactive orgs to be updated through HTTP API
Since LMS/Studio organizations have been completely backfilled,
there are now hundreds of Inactive organizations in the
edxapp database. Inactive organizations are not surfaced to
users, and hence are effectively non-existent.

If an organization is created or edited in an external
data source (such as Course Discovery), that source
should be able to push the update to the Organizations
API exposed by edx-organizations. Currently, however,
the Organizations API will return 400 if one tries to
create/update an organization that exists but is
inactive. This could potentially lead to Course
Discovery <-> LMS/Studio data synchronization issues
down the road.

This commit fixes that by making three changes:
 1. Any organization may be updated via the PUT method,
    regardless of whether or not it is active.
 2. Upon being updated, the organization is marked as
    active, whether it was previously active, inactive,
    or non-existent.
 3. The 'active' field may not be set or updated via
    the HTTP API, because that would be in conflict
    with change #2.

Bump version to 6.8.0.

TNL-7913